+++ src/libstddjb/bufalloc_init.c
@@ -4,7 +4,7 @@
 #include <skalibs/stralloc.h>
 #include <skalibs/bufalloc.h>
 
-void bufalloc_init (bufalloc *ba, int (*op)(int, char const *, size_t), int fd)
+void bufalloc_init (bufalloc *ba, ssize_t (*op)(int, char const *, size_t), int fd)
 {
   ba->x = stralloc_zero ;
   ba->op = op ;
