+++ src/libstddjb/cbuffer_init.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/cbuffer.h>
 
-int cbuffer_init (cbuffer_t *b, char *s, unsigned int len)
+int cbuffer_init (cbuffer_t *b, char *s, size_t len)
 {
   if (len < 2) return (errno = EINVAL, 0) ;
   b->x = s ;
