+++ src/libstddjb/pathexec_r.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/djbunix.h>
 
-void pathexec_r (char const *const *argv, char const *const *envp, unsigned int envlen, char const *modifs, unsigned int modiflen)
+void pathexec_r (char const *const *argv, char const *const *envp, size_t envlen, char const *modifs, size_t modiflen)
 {
   pathexec_r_name(argv[0], argv, envp, envlen, modifs, modiflen) ;
 }
