+++ src/libstddjb/rm_rf_tmp.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/djbunix.h>
 
 int rm_rf_tmp (char const *filename, stralloc *tmp)
 {
-  unsigned int tmpbase = tmp->len ;
+  size_t tmpbase = tmp->len ;
   if (!stralloc_cats(tmp, filename)) return -1 ;
   if (!stralloc_0(tmp)) goto err ;
   if (rm_rf_in_tmp(tmp, tmpbase) == -1) goto err ;
