+++ src/libstddjb/string_quote.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/skamisc.h>
 
-int string_quote (stralloc *sa, char const *s, unsigned int len)
+int string_quote (stralloc *sa, char const *s, size_t len)
 {
-  unsigned int base = sa->len ;
+  size_t base = sa->len ;
   int wasnull = !sa->s ;
   if (!stralloc_catb(sa, "\"", 1)) return 0 ;
   if (!string_quote_nodelim(sa, s, len) || !stralloc_catb(sa, "\"", 1))
