+++ src/libstddjb/vbprintf.c
@@ -1,5 +1,6 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <skalibs/buffer.h>
@@ -16,8 +17,8 @@ int vbprintf (buffer *b, char const *format, va_list args)
   }
   if (r < 0) return r ;
   {
-    char buf[(unsigned int)r + 1] ;
-    r = vsnprintf(buf, (unsigned int)r + 1, format, args) ;
+    char buf[(size_t)r + 1] ;
+    r = vsnprintf(buf, (size_t)r + 1, format, args) ;
     if (r < 0) return r ;
     if (buffer_put(b, buf, r) < r) return -1 ;
   }
